+++ Kmarket1/src/main/java/kr/co/kmarket1/controller/product/DeleteCartController.java
@@ -2,23 +2,17 @@ package kr.co.kmarket1.controller.product;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.ArrayList;
-import java.util.List;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 
 import com.google.gson.JsonObject;
 
-import kr.co.kmarket1.dao.ProductDAO;
-import kr.co.kmarket1.vo.ProductCartVO;
-import kr.co.kmarket1.vo.ProductVO;
+import kr.co.kmarket1.dao.ProductCartDAO;
 
 @WebServlet("/product/deleteCart.do")
 public class DeleteCartController extends HttpServlet{
@@ -36,7 +30,7 @@ public class DeleteCartController extends HttpServlet{
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		// ProductDAO 객체 생성
-		ProductDAO dao = ProductDAO.getInstance();
+		ProductCartDAO dao = ProductCartDAO.getInstance();
 		// DELETE_PRODUCT_CART 하는거임
 		String chks[] = req.getParameterValues("chk");
