+++ Kmarket1/src/main/java/kr/co/kmarket1/controller/product/OrderController.java
@@ -29,28 +29,38 @@ public class OrderController extends HttpServlet{
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		String prodNo = req.getParameter("prodNo");
 		String count = req.getParameter("count");
-		
-		ProductVO vo = ProductDAO.getInstance().selectOrderProduct(prodNo, count);
-		List<ProductVO> carts = new ArrayList<>();
-		carts.add(vo);
-		
-		req.setAttribute("carts", carts);
-		HttpSession session = req.getSession();
-		session.setAttribute("carts", carts);
+		if(prodNo != null) {
+			
+			ProductVO vo = ProductDAO.getInstance().selectOrderProduct(prodNo, count);
+			List<ProductVO> carts = new ArrayList<>();
+			carts.add(vo);
+			
+			req.setAttribute("carts", carts);
+			HttpSession session = req.getSession();
+			session.setAttribute("carts", carts);
 
-		
-		RequestDispatcher dispatcher = req.getRequestDispatcher("/product/order.jsp");
-		dispatcher.forward(req, resp);
+			
+			RequestDispatcher dispatcher = req.getRequestDispatcher("/product/order.jsp");
+			dispatcher.forward(req, resp);
+		}else {
+			HttpSession session = req.getSession();
+			List<ProductCartVO> carts = (List<ProductCartVO>) session.getAttribute("carts");
+			req.setAttribute("carts", carts);
+			
+			RequestDispatcher dispatcher = req.getRequestDispatcher("/product/order.jsp");
+			dispatcher.forward(req, resp);
+		}
 	}
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		HttpSession session = req.getSession();
-		@SuppressWarnings("unchecked")
-		List<ProductCartVO> carts = (List<ProductCartVO>) session.getAttribute("carts");
-		
-		req.setAttribute("carts", carts);
-		
-		RequestDispatcher dispatcher = req.getRequestDispatcher("/product/order.jsp");
-		dispatcher.forward(req, resp);
+// 여기부터 해야함		
+		// 주문번호 하나에 count는 중복된거 다
+		String prodNo = req.getParameter("prodNo");
+		String ordUid = req.getParameter("uid");
+		String recipName = req.getParameter("orderer");
+		String recipHp = req.getParameter("hp");
+		String recipZip = req.getParameter("zip");
+		String recipAddr1 = req.getParameter("addr1");
+		String recipAddr2 = req.getParameter("addr2");
 	}
 }
